Allow to specify policy on which nodes are used.
When creating EGLDisplay we can use DRM primary or render node, ie.
 - /dev/cardX, or
 - /dev/renderY

Now one can select which of those we use, and possibly to which we
should fallback too.

Please note:
Failure to open given node will at this moment treated as fatal
and fallback won't be attempted.

This is potentially subject to change, since this feature could be more
useful.